requirements.txt
discord.py==2.4.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
Flask>=3.0.0
waitress>=3.0.0
aiohttp>=3.9.5
discord.py==2.4.0
Flask>=3.0.0
python-dotenv>=1.0.1
waitress>=3.0.0
